Calculate Estimated Glomerular Filtration Rate for Children using the Schwartz Formula

Calculate Estimated Glomerular Filtration Rate (eGFR) for Pedatrics using the Schwartz formula

calculate_egfr_child(age, htcm, scr, sexf)

Arguments

age

Age (y). Numeric vector.

htcm

Height (cm). Numeric vector.

scr

Serum Creatinine (mg/dL). Numeric vector.

sexf

Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.

Value

This function returns a numeric vector the same length as its inputs

Details

Formula to calculate EGFR_CHILD:
  Children (AGE [y] <= 16), where:
  EGFR = (k * HTCM [cm]) / SCR [mg/dL]
  k = 0.45 for AGE [y] <= 1 (full term)
  k = 0.55 for girls 1 < AGE [y] <= 16
  k = 0.55 for boys 1 < AGE [y] < 13
  k = 0.70 for boys 13 <= AGE [y] <= 16

References

Schwartz GJ, Furth SL. Glomerular filtration rate measurement and estimation in chronic kidney disease. Pediatr Nephrol. 2007;22(11):1839-48.

See also

Examples

library(dplyr)

dmcognigen_cov %>% 
  mutate(EGFRSCHW = calculate_egfr_child(
    age = AGE, 
    htcm = HTCM, 
    scr = SCR, 
    sexf = SEXF
  ))
#> Formula to calculate EGFR_CHILD: 
#>   Children (AGE [y] <= 16), where:
#>   EGFR = (k * HTCM [cm]) / SCR [mg/dL]
#>   k = 0.45 for AGE [y] <= 1 (full term)
#>   k = 0.55 for girls 1 < AGE [y] <= 16
#>   k = 0.55 for boys 1 < AGE [y] < 13
#>   k = 0.70 for boys 13 <= AGE [y] <= 16
#> Warning: There was 1 warning in `mutate()`.
#> ℹ In argument: `EGFRSCHW = calculate_egfr_child(age = AGE, htcm = HTCM, scr =
#>   SCR, sexf = SEXF)`.
#> Caused by warning:
#> ! Cases where AGE > 16 were detected. This formula is intended for children.
#> # A tibble: 254 × 49
#>    STUDYID      DOMAIN USUBJID   SEXF SEXFC RACEN  HTCM  WTKG   AST ASTULN   SCR
#>    <chr>        <chr>  <chr>    <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#>  1 CDISCPILOT01 DM     01-701-…     1 Fema…     1  147.  54.4    40     34   0.9
#>  2 CDISCPILOT01 DM     01-701-…     0 Male      1  163.  80.3    21     36   1.4
#>  3 CDISCPILOT01 DM     01-701-…     0 Male      1  178.  99.3    24     36   1.4
#>  4 CDISCPILOT01 DM     01-701-…     0 Male      1  175.  88.4    20     36   1.5
#>  5 CDISCPILOT01 DM     01-701-…     1 Fema…     1  155.  62.6    23     34   1  
#>  6 CDISCPILOT01 DM     01-701-…     1 Fema…     1  149.  67.1    25     34   1  
#>  7 CDISCPILOT01 DM     01-701-…     0 Male      1  169.  78.0    19     36   1.4
#>  8 CDISCPILOT01 DM     01-701-…     1 Fema…     1  158.  59.9    28     34   0.9
#>  9 CDISCPILOT01 DM     01-701-…     0 Male      1  182.  78.9    26     36   1.3
#> 10 CDISCPILOT01 DM     01-701-…     0 Male      1  180.  71.2    15     36   1  
#> # ℹ 244 more rows
#> # ℹ 38 more variables: SCRULN <dbl>, TBIL <dbl>, TBILULN <dbl>, ASTCAT <dbl>,
#> #   BMI <dbl>, BSA <dbl>, IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>,
#> #   EGFRSCHW <dbl>, IBWCHILD <dbl>, LBM <dbl>, TBILCAT <dbl>, RFCAT <dbl>,
#> #   NCILIV <dbl>, SUBJID <chr>, RFSTDTC <chr>, RFENDTC <chr>, RFXSTDTC <chr>,
#> #   RFXENDTC <chr>, RFICDTC <chr>, RFPENDTC <chr>, DTHDTC <chr>, DTHFL <chr>,
#> #   SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>, ETHNIC <chr>, …

# Below will also work if the dataset contains expected variables
dmcognigen_cov %>% 
  mutate(EGFRSCHW = calculate_egfr_child())
#> AGE variable found and used for the age argument.
#> HTCM variable found and used for the htcm argument.
#> SCR variable found and used for the scr argument.
#> SEXF variable found and used for the sexf argument.
#> Formula to calculate EGFR_CHILD: 
#>   Children (AGE [y] <= 16), where:
#>   EGFR = (k * HTCM [cm]) / SCR [mg/dL]
#>   k = 0.45 for AGE [y] <= 1 (full term)
#>   k = 0.55 for girls 1 < AGE [y] <= 16
#>   k = 0.55 for boys 1 < AGE [y] < 13
#>   k = 0.70 for boys 13 <= AGE [y] <= 16
#> Warning: There was 1 warning in `mutate()`.
#> ℹ In argument: `EGFRSCHW = calculate_egfr_child()`.
#> Caused by warning:
#> ! Cases where AGE > 16 were detected. This formula is intended for children.
#> # A tibble: 254 × 49
#>    STUDYID      DOMAIN USUBJID   SEXF SEXFC RACEN  HTCM  WTKG   AST ASTULN   SCR
#>    <chr>        <chr>  <chr>    <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#>  1 CDISCPILOT01 DM     01-701-…     1 Fema…     1  147.  54.4    40     34   0.9
#>  2 CDISCPILOT01 DM     01-701-…     0 Male      1  163.  80.3    21     36   1.4
#>  3 CDISCPILOT01 DM     01-701-…     0 Male      1  178.  99.3    24     36   1.4
#>  4 CDISCPILOT01 DM     01-701-…     0 Male      1  175.  88.4    20     36   1.5
#>  5 CDISCPILOT01 DM     01-701-…     1 Fema…     1  155.  62.6    23     34   1  
#>  6 CDISCPILOT01 DM     01-701-…     1 Fema…     1  149.  67.1    25     34   1  
#>  7 CDISCPILOT01 DM     01-701-…     0 Male      1  169.  78.0    19     36   1.4
#>  8 CDISCPILOT01 DM     01-701-…     1 Fema…     1  158.  59.9    28     34   0.9
#>  9 CDISCPILOT01 DM     01-701-…     0 Male      1  182.  78.9    26     36   1.3
#> 10 CDISCPILOT01 DM     01-701-…     0 Male      1  180.  71.2    15     36   1  
#> # ℹ 244 more rows
#> # ℹ 38 more variables: SCRULN <dbl>, TBIL <dbl>, TBILULN <dbl>, ASTCAT <dbl>,
#> #   BMI <dbl>, BSA <dbl>, IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>,
#> #   EGFRSCHW <dbl>, IBWCHILD <dbl>, LBM <dbl>, TBILCAT <dbl>, RFCAT <dbl>,
#> #   NCILIV <dbl>, SUBJID <chr>, RFSTDTC <chr>, RFENDTC <chr>, RFXSTDTC <chr>,
#> #   RFXENDTC <chr>, RFICDTC <chr>, RFPENDTC <chr>, DTHDTC <chr>, DTHFL <chr>,
#> #   SITEID <chr>, AGE <dbl>, AGEU <chr>, SEX <chr>, RACE <chr>, ETHNIC <chr>, …