Calculate Lean Body Mass

Calculate Lean Body Mass (LBM), an estimation of how much someone weighs without the body fat; how much a persons bones, organs and muscles weigh

calculate_lbm(htcm, sexf, wtkg)

Arguments

htcm

Height (cm). Numeric vector.

sexf

Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.

wtkg

Weight (kg). Numeric vector.

Value

This function returns a numeric vector the same length as its inputs

Details

Formula to calculate LBM:
  Males: LBM [kg] = (1.10 × WTKG [kg]) - 128 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
  Females: LBM [kg] = (1.07 × WTKG [kg]) - 148 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)

References

Hallynck TH, Soep HH, Thomis JA, et al. Should clearance be normalised to body surface or to lean body mass? Br J Clin Pharmacol. 1981;11:523-526

James WPT. Research on obesity. London. Her Majesty's Stationery Office. 1976

Examples

library(dplyr)

dmcognigen_cov %>% 
  mutate(LBM = calculate_lbm(
    htcm = HTCM,
    sexf = SEXF, 
    wtkg = WTKG 
  ))
#> Formula to calculate LBM: 
#>   Males: LBM [kg] = (1.10 × WTKG [kg]) - 128 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
#>   Females: LBM [kg] = (1.07 × WTKG [kg]) - 148 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
#> # A tibble: 254 × 36
#>    STUDYID USUBJID SUBJID   AGE SEX    SEXF SEXFC RACE  RACEN ETHNIC ARMCD ARM  
#>    <chr>   <chr>   <chr>  <dbl> <chr> <dbl> <chr> <chr> <dbl> <chr>  <chr> <chr>
#>  1 CDISCP… 01-701… 1015      63 F         1 Fema… WHITE     1 HISPA… Pbo   Plac…
#>  2 CDISCP… 01-701… 1023      64 M         0 Male  WHITE     1 HISPA… Pbo   Plac…
#>  3 CDISCP… 01-701… 1028      71 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  4 CDISCP… 01-701… 1033      74 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  5 CDISCP… 01-701… 1034      77 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  6 CDISCP… 01-701… 1047      85 F         1 Fema… WHITE     1 NOT H… Pbo   Plac…
#>  7 CDISCP… 01-701… 1097      68 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  8 CDISCP… 01-701… 1111      81 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  9 CDISCP… 01-701… 1115      84 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#> 10 CDISCP… 01-701… 1118      52 M         0 Male  WHITE     1 NOT H… Pbo   Plac…
#> # ℹ 244 more rows
#> # ℹ 24 more variables: ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>, HTCM <dbl>,
#> #   WTKG <dbl>, AST <dbl>, ASTULN <dbl>, SCR <dbl>, SCRULN <dbl>, TBIL <dbl>,
#> #   TBILULN <dbl>, ASTCAT <dbl>, TBILCAT <dbl>, BMI <dbl>, BSA <dbl>,
#> #   IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>, EGFRSCHW <dbl>,
#> #   IBWCHILD <dbl>, LBM <dbl>, RFCAT <dbl>, NCILIV <dbl>

# Below will also work if the dataset contains expected variables
dmcognigen_cov %>% 
  mutate(LBM = calculate_lbm())
#> HTCM variable found and used for the htcm argument.
#> SEXF variable found and used for the sexf argument.
#> WTKG variable found and used for the wtkg argument.
#> Formula to calculate LBM: 
#>   Males: LBM [kg] = (1.10 × WTKG [kg]) - 128 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
#>   Females: LBM [kg] = (1.07 × WTKG [kg]) - 148 × ((WTKG [kg]) ^ 2 ÷ (HTCM [cm]) ^ 2)
#> # A tibble: 254 × 36
#>    STUDYID USUBJID SUBJID   AGE SEX    SEXF SEXFC RACE  RACEN ETHNIC ARMCD ARM  
#>    <chr>   <chr>   <chr>  <dbl> <chr> <dbl> <chr> <chr> <dbl> <chr>  <chr> <chr>
#>  1 CDISCP… 01-701… 1015      63 F         1 Fema… WHITE     1 HISPA… Pbo   Plac…
#>  2 CDISCP… 01-701… 1023      64 M         0 Male  WHITE     1 HISPA… Pbo   Plac…
#>  3 CDISCP… 01-701… 1028      71 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  4 CDISCP… 01-701… 1033      74 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  5 CDISCP… 01-701… 1034      77 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  6 CDISCP… 01-701… 1047      85 F         1 Fema… WHITE     1 NOT H… Pbo   Plac…
#>  7 CDISCP… 01-701… 1097      68 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  8 CDISCP… 01-701… 1111      81 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  9 CDISCP… 01-701… 1115      84 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#> 10 CDISCP… 01-701… 1118      52 M         0 Male  WHITE     1 NOT H… Pbo   Plac…
#> # ℹ 244 more rows
#> # ℹ 24 more variables: ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>, HTCM <dbl>,
#> #   WTKG <dbl>, AST <dbl>, ASTULN <dbl>, SCR <dbl>, SCRULN <dbl>, TBIL <dbl>,
#> #   TBILULN <dbl>, ASTCAT <dbl>, TBILCAT <dbl>, BMI <dbl>, BSA <dbl>,
#> #   IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>, EGFRSCHW <dbl>,
#> #   IBWCHILD <dbl>, LBM <dbl>, RFCAT <dbl>, NCILIV <dbl>