Count search results

When patterns are matched in variable names or labels, no subset is performed. When patterns are only matched in variable content, the data.frames are subset to rows where some variable matches the pattern.

cnt_search_result(
  search_result,
  n_distinct_vars = NULL,
  n_distinct_combined = TRUE,
  df_names = NULL,
  extra_vars = NULL,
  ignore_vars = NULL
)

Arguments

search_result

a search_result object obtained from search_environment_data.

n_distinct_vars

variables to count the unique values of.

n_distinct_combined

logical indicating whether to count the unique values of the combinations of n_distinct_vars.

df_names

an optional character vector of data.frame names to subset to.

extra_vars, ignore_vars

optional character vectors of variables to include/exclude in cnt results.

Value

a named list of cnt results.

See also

Examples

data(dmcognigen_cov)
study_search <- search_environment_data(
  stringr::regex("study", ignore_case = TRUE)
)
cnt_search_result(
  study_search, 
  n_distinct_vars = c("USUBJID", "SUBJID"), 
  extra_vars = "ACTARM"
)
#> $dmcognigen_cov
#> # A tibble: 3 × 6
#>   STUDYID      ACTARM               n_USUBJID n_SUBJID n_USUBJID_SUBJID     n
#>   <chr>        <chr>                    <int>    <int>            <int> <int>
#> 1 CDISCPILOT01 Placebo                     86       86               86    86
#> 2 CDISCPILOT01 Xanomeline High Dose        72       72               72    72
#> 3 CDISCPILOT01 Xanomeline Low Dose         96       96               96    96
#>