Example PK Data Requirements

Data Requirements designed based on pharmaversesdtm example datasets, which are updated SDTM datasets that use data from the CDISC pilot project.

This has the same structure and attributes as results from read_requirements and as_requirements.

Usage

dmcognigen_pk_requirements

Format

An object of class requirements (inherits from tbl_df, tbl, data.frame) with 54 rows and 5 columns.

Examples

head(dmcognigen_pk_requirements)
#> # A tibble: 6 × 5
#>   variable_name variable_label            pk_ard pk_mif format_decode
#>   <chr>         <chr>                     <chr>  <chr>  <chr>        
#> 1 ONUM          Overall Sequence Number   x      "x"    ""           
#> 2 NUM           Sequence Number           x      "x"    ""           
#> 3 STUDYID       Study Identifier          x      ""     ""           
#> 4 USUBJID       Unique Subject Identifier x      ""     ""           
#> 5 ID            Subject ID                x      "x"    ""           
#> 6 TSFD          Time Since First Dose (h) x      "x"    ""           

as_decode_tbls(dmcognigen_pk_requirements)
#> 
#> ── Decode tables ───────────────────────────────────────────────────────────────
#> 
#> ── DVID ──
#> 
#> 0=Dose
#> 1=Xanomeline Concentration (ug/mL)
#> 
#> ── EVID ──
#> 
#> 0=PK or PD measure
#> 1=Dose
#> 2=Other
#> 
#> ── MDV ──
#> 
#> 0=PK or PD measure
#> 1=Dose or Other
#> 
#> ── BLQFN ──
#> 
#> 0=No
#> 1=Yes
#> 
#> ── FED ──
#> 
#> 0=Fasted
#> 1=Fed
#> 
#> ── RACEN ──
#> 
#> 1=White/Caucasian
#> 2=Black/African American
#> 3=Asian
#> 4=American Indian or Alaska Native
#> 
#> ── SEXF ──
#> 
#> 0=Male
#> 1=Female
#> 
#> ── RFCAT ──
#> 
#> 1=Normal Function (>=90 mL/min)
#> 2=Mild Impairment (60-89 mL/min)
#> 3=Moderate Impairment (30-59 mL/min)
#> 4=Severe Impairment (15-29 mL/min)
#> 5=End Stage Disease (<15 mL/min or Dialysis)
#> 
#> ── NCILIV ──
#> 
#> 0=Normal Group A
#> 1=Mild Group B1
#> 2=Mild Group B2
#> 3=Moderate Group C
#> 4=Severe Group D