Extract decode tables from data

Extract decode tables from data

Usage

extract_decode_tbls_from_data(.data, lvl_to_lbl = "{var}C")

Arguments

.data

data frame

lvl_to_lbl

a named list, where elements correspond to matching names in .data and decode_tbls. One element can be left unnamed to provide the default transformation. Each element should be either: (1) a function/formula; (2) a custom glue specification. The default is "{var}C", which means the label variable name is the level variable name appended with a "C".

Value

decode_tbls

Examples

dmcognigen_cov |> 
  extract_decode_tbls_from_data()
#> 
#> ── Decode tables ───────────────────────────────────────────────────────────────
#> 
#> ── SEXF ──
#> 
#> 0=Male
#> 1=Female
#> 
#> ── RFCAT ──
#> 
#> 2=Mild Impairment (60-89 mL/min)
#> 3=Moderate Impairment (30-59 mL/min)
#> NA=NA
#> 
#> ── NCILIV ──
#> 
#> 0=Normal Group A
#> 1=Mild Group B1
#> 2=Mild Group B2
#> 3=Moderate Group C
#> 
#> ── RACE ──
#> 
#> AMERICAN INDIAN OR ALASKA NATIVE=American Indian or Alaska Native
#> BLACK OR AFRICAN AMERICAN=Black/African American
#> WHITE=White/Caucasian

dmcognigen_cov |> 
  extract_decode_tbls_from_data(
    lvl_to_lbl = c(
    # specify mapping from lvl to lbl by name
    "RACEN" = "RACE", 
    # one unnamed, default mapping can be provided
    ~ stringr::str_remove(.x, "CD$"))
  )
#> 
#> ── Decode tables ───────────────────────────────────────────────────────────────
#> 
#> ── RACEN ──
#> 
#> 1=WHITE
#> 2=BLACK OR AFRICAN AMERICAN
#> 4=AMERICAN INDIAN OR ALASKA NATIVE
#> 
#> ── ARMCD ──
#> 
#> Pbo=Placebo
#> Xan_Hi=Xanomeline High Dose
#> Xan_Lo=Xanomeline Low Dose
#> 
#> ── ACTARMCD ──
#> 
#> Pbo=Placebo
#> Xan_Hi=Xanomeline High Dose
#> Xan_Lo=Xanomeline Low Dose