Calculate Estimated Glomerular Filtration Rate (eGFR)

Calculate Estimated Glomerular Filtration Rate (eGFR) for adults

calculate_egfr(scr, age, sexf, racen)

Arguments

scr

Serum Creatinine (mg/dL). Numeric vector.

age

Age (y). Numeric vector.

sexf

Sex. Numeric vector including values of 0 and/or 1. 0=Male; 1=Female.

racen

Race. Numeric vector where the value 2 is expected to correspond to Black/African American.

Value

This function returns a numeric vector the same length as its inputs

Details

Formula to calculate EGFR: 
  EGFR [mL/min/1.73m^2] = 175 × (SCR [mg/dL] ^ -1.154) × (AGE [y] ^ -0.203) × (0.742 if female) × (1.212 if African American)

References

Guidance for Industry: Pharmacokinetics in Patients with Impaired Renal Function — Study Design, Data Analysis, and Impact on Dosing and Labeling

See also

Examples

library(dplyr)

dmcognigen_cov %>% 
  mutate(EGFR = calculate_egfr(
    scr = SCR, 
    age = AGE, 
    sexf = SEXF, 
    racen = RACEN
  ))
#> Formula to calculate EGFR: 
#>   EGFR [mL/min/1.73m^2] = 175 × (SCR [mg/dL] ^ -1.154) × (AGE [y] ^ -0.203) × (0.742 if female) × (1.212 if African American)
#> # A tibble: 254 × 36
#>    STUDYID USUBJID SUBJID   AGE SEX    SEXF SEXFC RACE  RACEN ETHNIC ARMCD ARM  
#>    <chr>   <chr>   <chr>  <dbl> <chr> <dbl> <chr> <chr> <dbl> <chr>  <chr> <chr>
#>  1 CDISCP… 01-701… 1015      63 F         1 Fema… WHITE     1 HISPA… Pbo   Plac…
#>  2 CDISCP… 01-701… 1023      64 M         0 Male  WHITE     1 HISPA… Pbo   Plac…
#>  3 CDISCP… 01-701… 1028      71 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  4 CDISCP… 01-701… 1033      74 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  5 CDISCP… 01-701… 1034      77 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  6 CDISCP… 01-701… 1047      85 F         1 Fema… WHITE     1 NOT H… Pbo   Plac…
#>  7 CDISCP… 01-701… 1097      68 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  8 CDISCP… 01-701… 1111      81 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  9 CDISCP… 01-701… 1115      84 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#> 10 CDISCP… 01-701… 1118      52 M         0 Male  WHITE     1 NOT H… Pbo   Plac…
#> # ℹ 244 more rows
#> # ℹ 24 more variables: ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>, HTCM <dbl>,
#> #   WTKG <dbl>, AST <dbl>, ASTULN <dbl>, SCR <dbl>, SCRULN <dbl>, TBIL <dbl>,
#> #   TBILULN <dbl>, ASTCAT <dbl>, TBILCAT <dbl>, BMI <dbl>, BSA <dbl>,
#> #   IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>, EGFRSCHW <dbl>,
#> #   IBWCHILD <dbl>, LBM <dbl>, RFCAT <dbl>, NCILIV <dbl>

# Below will also work if the dataset contains expected variables
dmcognigen_cov %>% 
  mutate(EGFR = calculate_egfr())
#> SCR variable found and used for the scr argument.
#> AGE variable found and used for the age argument.
#> SEXF variable found and used for the sexf argument.
#> RACEN variable found and used for the racen argument.
#> Formula to calculate EGFR: 
#>   EGFR [mL/min/1.73m^2] = 175 × (SCR [mg/dL] ^ -1.154) × (AGE [y] ^ -0.203) × (0.742 if female) × (1.212 if African American)
#> # A tibble: 254 × 36
#>    STUDYID USUBJID SUBJID   AGE SEX    SEXF SEXFC RACE  RACEN ETHNIC ARMCD ARM  
#>    <chr>   <chr>   <chr>  <dbl> <chr> <dbl> <chr> <chr> <dbl> <chr>  <chr> <chr>
#>  1 CDISCP… 01-701… 1015      63 F         1 Fema… WHITE     1 HISPA… Pbo   Plac…
#>  2 CDISCP… 01-701… 1023      64 M         0 Male  WHITE     1 HISPA… Pbo   Plac…
#>  3 CDISCP… 01-701… 1028      71 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  4 CDISCP… 01-701… 1033      74 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  5 CDISCP… 01-701… 1034      77 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  6 CDISCP… 01-701… 1047      85 F         1 Fema… WHITE     1 NOT H… Pbo   Plac…
#>  7 CDISCP… 01-701… 1097      68 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#>  8 CDISCP… 01-701… 1111      81 F         1 Fema… WHITE     1 NOT H… Xan_… Xano…
#>  9 CDISCP… 01-701… 1115      84 M         0 Male  WHITE     1 NOT H… Xan_… Xano…
#> 10 CDISCP… 01-701… 1118      52 M         0 Male  WHITE     1 NOT H… Pbo   Plac…
#> # ℹ 244 more rows
#> # ℹ 24 more variables: ACTARMCD <chr>, ACTARM <chr>, COUNTRY <chr>, HTCM <dbl>,
#> #   WTKG <dbl>, AST <dbl>, ASTULN <dbl>, SCR <dbl>, SCRULN <dbl>, TBIL <dbl>,
#> #   TBILULN <dbl>, ASTCAT <dbl>, TBILCAT <dbl>, BMI <dbl>, BSA <dbl>,
#> #   IBW <dbl>, CRCL <dbl>, CRCLP <dbl>, EGFR <dbl>, EGFRSCHW <dbl>,
#> #   IBWCHILD <dbl>, LBM <dbl>, RFCAT <dbl>, NCILIV <dbl>