Build decode tables

Objects of class decode_tbl include all pairs of lvl and lbl for a single var. They are generated based on variable definitions or values within a dataset.

Objects of decode_tbls are lists of decode_tbl objects.

as_decode_tbl(x = tibble::tibble())

as_decode_tbls(x = list())

Arguments

x

A tibble or data frame for as_decode_tbl(). A list for as_decode_tbls().

Value

a decode_tbl or decode_tbls object.

See also